Test auth uses Streamlit link button
